+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.5
+## 0.0.6+1
+RouteNode return route get-property with type AutoRoute
+RouteNodeCupertino return route get-property with type CupertinoRoute
+RouteNodeMaterial return route get-property with type MaterialRoute
 
+## 0.0.6
+RouteController is mixin now
+
+## 0.0.5
 My initial release.